LayoutTests/fast/dom/Range/range-exceptions-expected.txt

This test checks some DOM Range exceptions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS node.innerHTML is '<bar>AB<moo>C</moo>DE</bar>'
PASS r.surroundContents(document.createElement('a')) threw exception Error: InvalidStateError: DOM Exception 11.
PASS r.surroundContents(document.createElement('a')) threw exception Error: HierarchyRequestError: DOM Exception 3.
PASS successfullyParsed is true

TEST COMPLETE
